Refactor sync and participation field roots
