+++ beacon-chain/state/stateutil/BUILD.bazel
@@ -6,8 +6,10 @@ go_library(
         "array_root.go",
         "block_header_root.go",
         "eth1_root.go",
+        "participation_bit_root.go",
         "pending_attestation_root.go",
         "reference.go",
+        "sync_committee.root.go",
         "trie_helpers.go",
         "validator_map_handler.go",
         "validator_root.go",
