+++ beacon-chain/state/v2/BUILD.bazel
@@ -6,8 +6,6 @@ go_library(
         "deprecated_getters.go",
         "deprecated_setters.go",
         "field_root_eth1.go",
-        "field_root_participation_bit.go",
-        "field_root_sync_committee.go",
         "field_root_validator.go",
         "field_root_vector.go",
         "field_roots.go",
