+++ beacon-chain/state/v2/field_roots.go
@@ -163,14 +163,14 @@ func (h *stateRootHasher) computeFieldRootsWithHasher(ctx context.Context, state
 	fieldRoots[14] = slashingsRootsRoot[:]
 
 	// PreviousEpochParticipation slice root.
-	prevParticipationRoot, err := participationBitsRoot(state.PreviousEpochParticipation)
+	prevParticipationRoot, err := stateutil.ParticipationBitsRoot(state.PreviousEpochParticipation)
 	if err != nil {
 		return nil, errors.Wrap(err, "could not compute previous epoch participation merkleization")
 	}
 	fieldRoots[15] = prevParticipationRoot[:]
 
 	// CurrentEpochParticipation slice root.
-	currParticipationRoot, err := participationBitsRoot(state.CurrentEpochParticipation)
+	currParticipationRoot, err := stateutil.ParticipationBitsRoot(state.CurrentEpochParticipation)
 	if err != nil {
 		return nil, errors.Wrap(err, "could not compute current epoch participation merkleization")
 	}
@@ -209,14 +209,14 @@ func (h *stateRootHasher) computeFieldRootsWithHasher(ctx context.Context, state
 	fieldRoots[21] = inactivityScoresRoot[:]
 
 	// Current sync committee root.
-	currentSyncCommitteeRoot, err := syncCommitteeRoot(state.CurrentSyncCommittee)
+	currentSyncCommitteeRoot, err := stateutil.SyncCommitteeRoot(state.CurrentSyncCommittee)
 	if err != nil {
 		return nil, errors.Wrap(err, "could not compute sync committee merkleization")
 	}
 	fieldRoots[22] = currentSyncCommitteeRoot[:]
 
 	// Next sync committee root.
-	nextSyncCommitteeRoot, err := syncCommitteeRoot(state.NextSyncCommittee)
+	nextSyncCommitteeRoot, err := stateutil.SyncCommitteeRoot(state.NextSyncCommittee)
 	if err != nil {
 		return nil, errors.Wrap(err, "could not compute sync committee merkleization")
 	}
